+++ src/graphql/index.ts
@@ -1,0 +1,2 @@
+export { GraphQLResource } from './graphql.resource';
+export * from './graphql.types';